+++ v3/v3_structs.go
@@ -3,7 +3,7 @@ package v3
 import (
 	"time"
 
-	client "github.com/nutanix-cloud-native/prism-go-client"
+	"github.com/nutanix-cloud-native/prism-go-client"
 )
 
 // Reference ...
@@ -582,7 +582,7 @@ type DSMetadata struct {
 	SortOrder *string `json:"sort_order,omitempty" mapstructure:"sort_order,omitempty"`
 
 	// Additional filters for client side filtering api response
-	ClientSideFilters []*client.AdditionalFilter `json:"-"`
+	ClientSideFilters []*prismgoclient.AdditionalFilter `json:"-"`
 }
 
 // VMIntentResource Response object for intentful operations on a vm